daemon vérifie uptime et date modif

+++ GrenobleLiveWallpaper.py
@@ -2,17 +2,21 @@ import subprocess
 import os, signal
 import platform
 _platform = platform.system()
+import time, uptime
 
 if os.path.exists('.pid'):
-	c = input("Grenoble Live Wallpaper is already running. Do you wish to kill it? ([y]/n) ")
-	if c in ['', 'y', 'Y']:
-		with open('.pid', 'r') as pidfile:
-			pid = int(pidfile.readline())
-			print(pid)
-			if _platform == 'Windows':
-				subprocess.call(['taskkill', '/F', '/T', '/PID',  str(pid)])
-			else:
-				os.kill(pid, signal.SIGTERM)
+	if time.time() - uptime.uptime() < os.path.getmtime('.pid') : # si le fichier a été modifié après le démarrage de l'OS
+		c = input("Grenoble Live Wallpaper is already running. Do you wish to kill it? ([y]/n) ")
+		if c in ['', 'y', 'Y']:
+			with open('.pid', 'r') as pidfile:
+				pid = int(pidfile.readline())
+				print(pid)
+				if _platform == 'Windows':
+					subprocess.call(['taskkill', '/F', '/T', '/PID',  str(pid)])
+				else:
+					os.kill(pid, signal.SIGTERM)
+			os.remove('.pid')
+	else :
 		os.remove('.pid')
 
 else :
